+++ src/test/graph_test.go
@@ -16,23 +16,23 @@ func TestNewGraphFromAdjacencyMatrix(t *testing.T) {
 	adjacencyMatrix := models.NewAdjacencyMatrix(nodeCount, columnLabels, latitudes, longitudes)
 
 	adjacencyMatrix.Matrix[0][0] = 0
-	adjacencyMatrix.Matrix[0][1] = 1
-	adjacencyMatrix.Matrix[0][2] = 2
+	adjacencyMatrix.Matrix[0][1] = 1.0
+	adjacencyMatrix.Matrix[0][2] = 2.0
 	adjacencyMatrix.Matrix[0][3] = 0
 
-	adjacencyMatrix.Matrix[1][0] = 1
+	adjacencyMatrix.Matrix[1][0] = 1.0
 	adjacencyMatrix.Matrix[1][1] = 0
-	adjacencyMatrix.Matrix[1][2] = 3
+	adjacencyMatrix.Matrix[1][2] = 3.0
 	adjacencyMatrix.Matrix[1][3] = 0
 
-	adjacencyMatrix.Matrix[2][0] = 2
-	adjacencyMatrix.Matrix[2][1] = 3
+	adjacencyMatrix.Matrix[2][0] = 2.0
+	adjacencyMatrix.Matrix[2][1] = 3.0
 	adjacencyMatrix.Matrix[2][2] = 0
-	adjacencyMatrix.Matrix[2][3] = 4
+	adjacencyMatrix.Matrix[2][3] = 4.0
 
 	adjacencyMatrix.Matrix[3][0] = 0
 	adjacencyMatrix.Matrix[3][1] = 0
-	adjacencyMatrix.Matrix[3][2] = 4
+	adjacencyMatrix.Matrix[3][2] = 4.0
 	adjacencyMatrix.Matrix[3][3] = 0
 
 	graph := models.NewGraphFromAdjacencyMatrix(adjacencyMatrix)
@@ -52,11 +52,11 @@ func TestNewGraphFromAdjacencyMatrix(t *testing.T) {
 		}
 	}
 
-	expectedEdges := map[int]map[int]int64{
-		0: {1: 1, 2: 2},
-		1: {0: 1, 2: 3},
-		2: {0: 2, 1: 3, 3: 4},
-		3: {2: 4},
+	expectedEdges := map[int]map[int]float64{
+		0: {1: 1.0, 2: 2.0},
+		1: {0: 1.0, 2: 3.0},
+		2: {0: 2.0, 1: 3.0, 3: 4.0},
+		3: {2: 4.0},
 	}
 
 	if !reflect.DeepEqual(graph.Edges, expectedEdges) {
